// src/app/model/monthlyClosurePerson.ts
export class MonthlyClosurePerson {
    id?: number;
    personName: string;
    creditSalary: number;
    creditBenefit: number;
    creditOther: number;
    creditFull: number;
    debitFull: number;
    payable: number;
    percentage: number;
    payableMarket: number;
    percentageMarket: number;
    remaining: number;
    leisurePercent: number;
    leisureValue: number;
    investmentPercent: number;
    investmentValue: number;
}